requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
django-browser-reload==1.10.0
djangorestframework==3.14.0
environs==9.5.0
idna==3.4
marshmallow==3.19.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
whitenoise==6.4.0
